+++ changelog.md
@@ -15,7 +15,8 @@
 * Removed boolean `useAccel` in support of new mode paradigm. To use the accel simply set `curAccelMode` to `ACCEL_MODE_ON` or a function call `useAccel()`. You can also turn off the excel with a call to `useAccel(false)`.
 * Removed `OPENBCI_EOP_STND_ACCEL`, `OPENBCI_EOP_STND_RAW_AUX`, `OPENBCI_EOP_USER_DEFINED`, `OPENBCI_EOP_ACCEL_TIME_SET`, `OPENBCI_EOP_ACCEL_TIME_SYNCED`, `OPENBCI_EOP_RAW_AUX_TIME_SET`, and `OPENBCI_EOP_RAW_AUX_TIME_SYNCED` in place of `enum` support replacement: `PACKET_TYPE_ACCEL`, `PACKET_TYPE_RAW_AUX`, `PACKET_TYPE_USER_DEFINED`, `PACKET_TYPE_ACCEL_TIME_SET`, `PACKET_TYPE_ACCEL_TIME_SYNC`, `PACKET_TYPE_RAW_AUX_TIME_SET`, and `PACKET_TYPE_RAW_AUX_TIME_SYNC`.
 * Removed `sniffMode` in place for `curBoardMode == OPENBCI_BOARD_MODE_DEBUG`
-* Removed `sendChannelData(void)` function, use `sendChannelData(PACKET_TYPE)` and specify packet type. 
+* Removed `sendChannelData(void)` function, use `sendChannelData(PACKET_TYPE)` and specify packet type.
+* Removed public `waitForNewChannelData()` 
 
 # v2.0.1